requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
aif360>=0.5.0
fairlearn>=0.9.0
google-generativeai>=0.3.0
python-multipart>=0.0.6
pydantic>=2.0.0
